updated to 64bit

+++ src/cpus.rs
@@ -6,5 +6,5 @@ pub mod mc680x0;
 type Mnemonic = &'static str;
 
 pub trait CpuTrait : DisassemblerTrait {
-    fn set_program_counter(&mut self, pc:u32) -> Result<i32,i32>;
+    fn set_program_counter(&mut self, pc:u64) -> Result<i32,i32>;
 }
